add note controller

+++ src/main/java/ru/rabiarill/controllers/v1/NoteController.java
@@ -10,7 +10,9 @@ import ru.rabiarill.exception.model.NoAccessException;
 import ru.rabiarill.exception.model.note.NotValidNoteException;
 import ru.rabiarill.exception.model.note.NoteNotFoundException;
 import ru.rabiarill.models.note.Note;
+import ru.rabiarill.models.user.User;
 import ru.rabiarill.services.NoteService;
+import ru.rabiarill.util.security.UserUtil;
 
 import javax.validation.Valid;
 import java.util.List;
@@ -20,10 +22,12 @@ import java.util.List;
 public class NoteController {
 
    private final NoteService noteService;
+   private final UserUtil userUtil;
 
    @Autowired
-   public NoteController(NoteService noteService) {
+   public NoteController(NoteService noteService, UserUtil userUtil) {
       this.noteService = noteService;
+      this.userUtil = userUtil;
    }
 
    @GetMapping()
@@ -34,14 +38,33 @@ public class NoteController {
       return new ResponseEntity<>(response, HttpStatus.OK);
    }
 
-   @PutMapping("/{id}")
-   public ResponseEntity<HttpStatus> update(@RequestBody @Valid NoteDTO noteDTO,
-                                            BindingResult bindingResult,
-                                            @PathVariable("id") int id) throws NoAccessException, NoteNotFoundException {
+   @PostMapping()
+   public ResponseEntity<HttpStatus> create(@RequestBody @Valid NoteDTO noteDTO,
+                                            BindingResult bindingResult) {
       if (bindingResult.hasErrors())
          throw new NotValidNoteException();
 
+      Note noteToSave = noteDTO.convertToNote();
+
+      noteService.save(noteToSave);
+
+      return new ResponseEntity<>(HttpStatus.CREATED);
+   }
+
+   @PutMapping("/{id}")
+   public ResponseEntity<HttpStatus> update(@RequestBody @Valid NoteDTO noteDTO,
+                                            BindingResult bindingResult,
+                                            @PathVariable("id") int id) throws NoteNotFoundException {
+      if (bindingResult.hasErrors())
+         throw new NotValidNoteException();
+
+      Note noteDB = noteService.findOne(id);
+      User sender = userUtil.getUserFromContextHolder();
+      if (!userUtil.hasAccess(sender, noteDB))
+         throw new NoAccessException("You should be owner of note or has role \"ADMIN\"");
+
       Note noteToUpdate = noteDTO.convertToNote();
+      noteToUpdate.setOwner(noteDB.getOwner());
 
       noteService.update(noteToUpdate, id);
 
@@ -49,8 +72,13 @@ public class NoteController {
    }
 
    @DeleteMapping("/{id}")
-   public ResponseEntity<HttpStatus> deleteById(@PathVariable(value = "id") int id)
-           throws NoteNotFoundException, NoAccessException {
+   public ResponseEntity<HttpStatus> deleteById(@PathVariable(value = "id") int id) throws NoteNotFoundException {
+
+      User sender = userUtil.getUserFromContextHolder();
+      Note note  = noteService.findOne(id);
+
+      if (!userUtil.hasAccess(sender, note))
+         throw new NoAccessException("You should be owner of note or has role \"ADMIN\"");
 
       noteService.delete(id);
       return new ResponseEntity<>(HttpStatus.NO_CONTENT);
